Exploded tests into multiple files.
