+++ package.json
@@ -10,7 +10,7 @@
     "pino": "./pretty.js"
   },
   "scripts": {
-    "test": "standard && tap test.js"
+    "test": "standard && tap --cov test/*test.js"
   },
   "precommit": "test",
   "repository": {
